// src/lib/enums.ts
export enum AnimationType {
  SPIN = "spinEffect",
  BOUNCE = "bounceEffect",
  PULSE_AND_SCALE = "pulseAndScaleEffect",
}

export enum Colors {
  PRIMARY_COLOR = "var(--primary-color)",
  ORANGE_COLOR = "var(--orange-color)",
  GREEN_COLOR = "var(--green-color)",
  RED_COLOR = "var(--red-color)",
  DARK_BLUE_COLOR = "var(--dark-blue-color)",
  BACKGROUND_COLOR = "var(--background-color)",
  OPPOSITE_BACKGROUND_COLOR = "var(--opposite-background-color)",
}